% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R, R/tsoa_tmp.R
\name{auroc}
\alias{auroc}
\title{auroc}
\usage{
auroc(score, bool)

auroc(score, bool)
}
\arguments{
\item{score}{Score for ranking the observations.}

\item{bool}{Boolean giving the observation class.}
}
\value{
ROC AUC value

ROC AUC value
}
\description{
Fast function for calculating ROC AUC (original code by Miron Kursa https://mbq.me).

Fast function for calculating ROC AUC (original code by Miron Kursa https://mbq.me).
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
