% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R, R/tsoa_tmp.R
\name{calculate_autocorrelation}
\alias{calculate_autocorrelation}
\title{calculate_autocorrelation}
\usage{
calculate_autocorrelation(this_timeseries, lag = 1)

calculate_autocorrelation(this_timeseries, lag = 1)
}
\arguments{
\item{this_timeseries}{Measurements of a time series for an individual subject ranked by time.}

\item{lag}{Lag parameter for auto-correlation. Default lag is one.}
}
\value{
Autocorrelation coefficient.

Autocorrelation coefficient.
}
\description{
Calculates the autocorrelation feature for a time series. Default lag is one time point.

Calculates the autocorrelation feature for a time series. Default lag is one time point.
}
