% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctas.R
\name{calculate_entity_bias_ts_features}
\alias{calculate_entity_bias_ts_features}
\title{calculate_entity_bias_ts_features}
\usage{
calculate_entity_bias_ts_features(
  this_feature,
  this_data,
  this_ref_group,
  this_entity_level
)
}
\arguments{
\item{this_feature}{Type of the time series feature.}

\item{this_data}{Feature values for each subject.}

\item{this_ref_group}{Reference group for the entities (country, region or global).}

\item{this_entity_level}{Whether p-values should be calculated for sites, countries or regions.}
}
\value{
Data frame with KS p-values for each entity.
}
\description{
Performs Kolmogorov-Smirnov (KS) test for a particular parameter and time series feature.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
