% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctas.R
\name{calculate_lof}
\alias{calculate_lof}
\title{calculate_lof}
\usage{
calculate_lof(this_origvalues_distances)
}
\arguments{
\item{this_origvalues_distances}{Dist object for the time series}
}
\value{
Data frame with a feature value per subject.
}
\description{
Calculates Local Outlier Factor (lof) for each time series
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
