% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{calculate_own_site_simil_score}
\alias{calculate_own_site_simil_score}
\title{calculate_own_site_simil_score}
\usage{
calculate_own_site_simil_score(this_timeseries_wide)
}
\arguments{
\item{this_timeseries_wide}{Time series data in wide format.}
}
\value{
Data frame with a feature value per subject.
}
\description{
Calculates the own site similarity feature for each time series.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
