% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{calculate_principal_components}
\alias{calculate_principal_components}
\title{calculate_principal_components}
\usage{
calculate_principal_components(this_timeseries_wide)
}
\arguments{
\item{this_timeseries_wide}{Wide-format data frame with data to calculate the principal components for.}
}
\value{
Data frame with the principal components.
}
\description{
Calculate the first two principal components. These are used to visualize the relative distance between the time series/subjects.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
