% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{calculate_site_bias_ts_features}
\alias{calculate_site_bias_ts_features}
\title{calculate_site_bias_ts_features}
\usage{
calculate_site_bias_ts_features(this_feature, this_data)
}
\arguments{
\item{this_feature}{Type of the time series feature.}

\item{this_data}{Feature values for each subject.}
}
\value{
Data frame with KS p-values for each site.
}
\description{
Performs Kolmogorov-Smirnov (KS) test for a particular parameter and time series feature.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
