% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{calculate_ts_features}
\alias{calculate_ts_features}
\title{calculate_ts_features}
\usage{
calculate_ts_features(
  this_timeseries_wide,
  this_baseline,
  this_timeseries_features_to_calculate,
  subjects
)
}
\arguments{
\item{this_timeseries_wide}{Time series data in wide format.}

\item{this_baseline}{Whether the data in this_timeseries_wide are original or baseline-adjusted measurements.}

\item{this_timeseries_features_to_calculate}{Vector of time series features to return.}

\item{subjects}{Data frame with one row per study subject.}
}
\value{
Data frame with time series features.
}
\description{
Calculates features for a time series.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
