% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{check_input_data}
\alias{check_input_data}
\title{check_input_data}
\usage{
check_input_data()
}
\value{
no data returned
}
\description{
Makes sure the input data is in correct format etc. No return value is given. The function stops processing if something is not right.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
