% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctas_data}
\alias{ctas_data}
\title{Simulated input data to pass to process_a_study}
\format{
a list containing several dataframes
}
\usage{
ctas_data
}
\description{
Simulated input data to pass to process_a_study
}
\keyword{datasets}
