% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctas.R
\name{fit_mixed_effects_model}
\alias{fit_mixed_effects_model}
\title{fit_mixed_effects_model}
\usage{
fit_mixed_effects_model(this_data)
}
\arguments{
\item{this_data}{Feature values for each subject.}
}
\value{
Data frame with mixed effect model results for each entity (site, country and region)
}
\description{
Fits a mixed effects model to the data to identify any bias in time series feature bias for sites, countries and regions.
}
