% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{generate_wide_timeseries_table}
\alias{generate_wide_timeseries_table}
\title{generate_wide_timeseries_table}
\usage{
generate_wide_timeseries_table(
  this_parameter_id,
  this_timepoints,
  this_subjects,
  this_baseline,
  data
)
}
\arguments{
\item{this_parameter_id}{Parameter ID.}

\item{this_timepoints}{Semicolon-delimited string of time series timepoints.}

\item{this_subjects}{Semicolon-delimited string of subject IDs.}

\item{this_baseline}{If TRUE, the values should be baseline-adjusted.}

\item{data}{Measurements of study parameters. One row per data point.}
}
\value{
Time series data in wide format.
}
\description{
Transforms data for a time series as a wide data frame.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
