% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{get_max_sites_and_subjects}
\alias{get_max_sites_and_subjects}
\title{Get timpepoint that has the most sites and patients}
\usage{
get_max_sites_and_subjects(
  timepoint_ranks,
  this_time_point_count_min,
  this_max_share_missing,
  dataset,
  this_subject_count_min,
  subjects
)
}
\arguments{
\item{timepoint_ranks}{a vector with all time point ranks}

\item{this_time_point_count_min}{Minimum number of time points a time series must have.}

\item{this_max_share_missing}{Maximum share of missing measuremnents a subject can have.}

\item{dataset}{Description of parameter x.}

\item{this_subject_count_min}{Minimum number of subjects per time series.}

\item{subjects}{Data frame with one row per study subject.}
}
\value{
integer with optimized timepoint
}
\description{
Get timpepoint that has the most sites and patients
}
\details{
is called by pick_timepoint combos. uses the same for loop to iterate
over possible timeseries and returns eligable patients and sites. returns
highest timepoint rank that has the highest number of sites with the highest number of
patients. As NA values can occur randomly patient elegibility can vary depending on
the timepoint range. It is not necessarily the shortest range that has the most
patients. This is why we iteratively need to determine the optimal conditions
using the for loop.
}
