% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctas.R
\name{parse_readable_timeseries_combo_string}
\alias{parse_readable_timeseries_combo_string}
\title{parse_readable_timeseries_combo_string}
\usage{
parse_readable_timeseries_combo_string(
  timeseries_combo,
  this_parameter_id,
  this_timepoint_rank_to_name_mapping
)
}
\arguments{
\item{timeseries_combo}{Semicolon-delimited string with time point codes.}

\item{this_parameter_id}{Parameter the time point combo is for.}

\item{this_timepoint_rank_to_name_mapping}{Mapping table from time point code to human-readable name.}
}
\value{
Semicolon-delimited string of human-readable timepoint names.
}
\description{
Parse timepoint codes into a more human-readable format.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
