% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{pick_subjects_for_custom_timeseries}
\alias{pick_subjects_for_custom_timeseries}
\title{pick_subjects_for_custom_timeseries}
\usage{
pick_subjects_for_custom_timeseries(
  this_timepoints_and_subjects,
  this_timepoints_string,
  this_max_share_missing,
  this_parameter_id,
  this_baseline
)
}
\arguments{
\item{this_timepoints_and_subjects}{Data frame which tells which subjects have a result at which time point.}

\item{this_timepoints_string}{Time points that belong to the time series.}

\item{this_max_share_missing}{Maximum share of missing data points a subject can have.}

\item{this_parameter_id}{ID of the time series parameter.}

\item{this_baseline}{Whether the time series is for original or baseline-adjusted results.}
}
\value{
Semicolon-delimited string of subject IDs.
}
\description{
Picks subjects which have enough data for a time series.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
