% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoa.R
\name{pick_timepoint_combos}
\alias{pick_timepoint_combos}
\title{pick_timepoint_combos}
\usage{
pick_timepoint_combos(
  dataset,
  this_time_point_count_min,
  this_subject_count_min,
  this_max_share_missing,
  this_baseline
)
}
\arguments{
\item{dataset}{Description of parameter x.}

\item{this_time_point_count_min}{Minimum number of time points a time series must have.}

\item{this_subject_count_min}{Minimum number of subjects per time series.}

\item{this_max_share_missing}{Maximum share of missing measuremnents a subject can have.}

\item{this_baseline}{Whether the time series is for actual measurements or change-from-baseline values.}
}
\value{
Data frame with one or more time series for the parameter.
}
\description{
Autogenerates one or more time series for the parameter.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
