% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctas.R
\docType{data}
\name{tsoa_data}
\alias{tsoa_data}
\title{Simulated input data to pass to process_a_study}
\format{
a list containing several dataframes
}
\usage{
tsoa_data
}
\description{
Simulated input data to pass to process_a_study
}
\keyword{datasets}
