% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctas.R
\name{pick_timepoint_combos}
\alias{pick_timepoint_combos}
\title{pick_timepoint_combos}
\usage{
pick_timepoint_combos(
  dataset,
  this_time_point_count_min,
  this_subject_count_min,
  this_max_share_missing,
  this_baseline,
  subjects,
  optimize_sites_and_patients
)
}
\arguments{
\item{autogenerate_timeseries_type}{Defines the type of timeseries to be generated.}
\item{dataset}{Description of parameter x.}

\item{this_time_point_count_min}{Minimum number of time points a time series must have.}

\item{this_subject_count_min}{Minimum number of subjects per time series.}

\item{this_max_share_missing}{Maximum share of missing measuremnents a subject can have.}

\item{this_baseline}{Whether the time series is for actual measurements or change-from-baseline values.}

\item{subjects}{Data frame with one row per study subject.}

\item{optimize_sites_and_patients}{If set to TRUE, always creates timeseries with as many sites and patients as possible while respecting the other function parameters. Default:FALSE}
}
\value{
Data frame with one or more time series for the parameter.
}
\description{
Autogenerates one or more time series for the parameter.
}
\author{
Pekka Tiikkainen, \email{pekka.tiikkainen@bayer.com}
}
